refactor: move code to app module for better readability
