+++ src/main.rs
@@ -1,3 +1,4 @@
+pub mod app;
 pub mod db;
 pub mod error;
 pub mod flow;
@@ -7,26 +8,16 @@ pub mod models;
 pub mod scheduler;
 pub mod util;
 
+use crate::app::run;
+
+use anyhow::Result;
 use dotenvy::dotenv;
-use error::MostroError;
-use lightning::invoice::is_valid_invoice;
-use log::{error, info};
-
+use lightning::LndConnector;
 use nostr_sdk::prelude::*;
-
-use mostro_core::order::Order;
-use mostro_core::{order::SmallOrder, Action, Content, Message, Peer, Status};
 use scheduler::start_scheduler;
-use sqlx_crud::Crud;
-use std::str::FromStr;
-use tokio::sync::mpsc::channel;
-use tonic_openssl_lnd::lnrpc::payment::PaymentStatus;
-use util::*;
-
-use crate::db::update_order_to_initial_state;
 
 #[tokio::main]
-async fn main() -> anyhow::Result<()> {
+async fn main() -> Result<()> {
     dotenv().ok();
     pretty_env_logger::init();
     // Connect to database
@@ -40,673 +31,12 @@ async fn main() -> anyhow::Result<()> {
         .since(Timestamp::now());
 
     client.subscribe(vec![subscription]).await;
-    let mut ln_client = lightning::LndConnector::new().await;
+    let mut ln_client = LndConnector::new().await;
 
     // Start scheduler for tasks
     start_scheduler().await.unwrap().start().await?;
 
-    loop {
-        let mut notifications = client.notifications();
-
-        while let Ok(notification) = notifications.recv().await {
-            if let RelayPoolNotification::Event(_, event) = notification {
-                if let Kind::EncryptedDirectMessage = event.kind {
-                    let message = decrypt(
-                        &my_keys.secret_key().unwrap(),
-                        &event.pubkey,
-                        &event.content,
-                    );
-                    if let Ok(m) = message {
-                        let message = Message::from_json(&m);
-                        if let Ok(msg) = message {
-                            if msg.verify() {
-                                match msg.action {
-                                    Action::Order => {
-                                        if let Some(order) = msg.get_order() {
-                                            publish_order(
-                                                &pool,
-                                                &client,
-                                                &my_keys,
-                                                order,
-                                                &event.pubkey.to_bech32()?,
-                                            )
-                                            .await?
-                                        }
-                                    }
-                                    Action::TakeSell => {
-                                        // Safe unwrap as we verified the message
-                                        let order_id = msg.order_id.unwrap();
-                                        let mut order = match Order::by_id(&pool, order_id).await? {
-                                            Some(order) => order,
-                                            None => {
-                                                error!("TakeSell: Order Id {order_id} not found!");
-                                                break;
-                                            }
-                                        };
-                                        if order.kind != "Sell" {
-                                            error!("TakeSell: Order Id {order_id} wrong kind");
-                                            break;
-                                        }
-                                        let buyer_pubkey = event.pubkey;
-                                        let pr: Option<String>;
-                                        // If a buyer sent me a lightning invoice we look on db an order with
-                                        // that order id and save the buyer pubkey and invoice fields
-                                        if let Some(payment_request) = msg.get_payment_request() {
-                                            let order_amount = if order.amount == 0 {
-                                                None
-                                            } else {
-                                                Some(order.amount as u64)
-                                            };
-
-                                            // Verify if invoice is valid
-                                            match is_valid_invoice(&payment_request, order_amount) {
-                                                Ok(_) => {}
-                                                Err(e) => match e {
-                                                    MostroError::ParsingInvoiceError
-                                                    | MostroError::InvoiceExpiredError
-                                                    | MostroError::MinExpirationTimeError
-                                                    | MostroError::WrongAmountError
-                                                    | MostroError::MinAmountError => {
-                                                        send_dm(
-                                                            &client,
-                                                            &my_keys,
-                                                            &buyer_pubkey,
-                                                            e.to_string(),
-                                                        )
-                                                        .await?;
-                                                        error!("{e}");
-                                                        break;
-                                                    }
-                                                    _ => {}
-                                                },
-                                            }
-                                            pr = Some(payment_request);
-                                        } else {
-                                            pr = None;
-                                        }
-
-                                        let order_status = match Status::from_str(&order.status) {
-                                            Ok(s) => s,
-                                            Err(e) => {
-                                                error!("TakeSell: Order Id {order_id} wrong status: {e:?}");
-                                                break;
-                                            }
-                                        };
-                                        // Buyer can take pending orders only
-                                        match order_status {
-                                            Status::Pending | Status::WaitingBuyerInvoice => {}
-                                            _ => {
-                                                send_dm(
-                                                    &client,
-                                                    &my_keys,
-                                                    &buyer_pubkey,
-                                                    format!(
-                                                        "Order Id {order_id} was already taken!"
-                                                    ),
-                                                )
-                                                .await?;
-                                                break;
-                                            }
-                                        }
-                                        let seller_pubkey = match order.seller_pubkey.as_ref() {
-                                            Some(pk) => XOnlyPublicKey::from_bech32(pk)?,
-                                            None => {
-                                                error!(
-                                                    "TakeSell: Seller pubkey not found for order {}!",
-                                                    order.id
-                                                );
-                                                break;
-                                            }
-                                        };
-
-                                        // Check market price value in sats - if order was with market price then calculate it and send a DM to buyer
-                                        if order.amount == 0 {
-                                            order.amount = set_market_order_sats_amount(
-                                                &mut order,
-                                                buyer_pubkey,
-                                                &my_keys,
-                                                &pool,
-                                                &client,
-                                            )
-                                            .await?;
-                                        } else {
-                                            show_hold_invoice(
-                                                &pool,
-                                                &client,
-                                                &my_keys,
-                                                pr,
-                                                &buyer_pubkey,
-                                                &seller_pubkey,
-                                                &order,
-                                            )
-                                            .await?;
-                                        }
-                                    }
-                                    Action::TakeBuy => {
-                                        let seller_pubkey = event.pubkey;
-                                        // Safe unwrap as we verified the message
-                                        let order_id = msg.order_id.unwrap();
-                                        let mut order = match Order::by_id(&pool, order_id).await? {
-                                            Some(order) => order,
-                                            None => {
-                                                error!("TakeBuy: Order Id {order_id} not found!");
-                                                break;
-                                            }
-                                        };
-                                        if order.kind != "Buy" {
-                                            error!("TakeBuy: Order Id {order_id} wrong kind");
-                                            break;
-                                        }
-
-                                        let order_status = match Status::from_str(&order.status) {
-                                            Ok(s) => s,
-                                            Err(e) => {
-                                                error!("TakeBuy: Order Id {order_id} wrong status: {e:?}");
-                                                break;
-                                            }
-                                        };
-                                        // Buyer can take pending orders only
-                                        if order_status != Status::Pending {
-                                            send_dm(
-                                                &client,
-                                                &my_keys,
-                                                &seller_pubkey,
-                                                format!("Order Id {order_id} was already taken!"),
-                                            )
-                                            .await?;
-                                            break;
-                                        }
-                                        let buyer_pubkey = match order.buyer_pubkey.as_ref() {
-                                            Some(pk) => XOnlyPublicKey::from_bech32(pk)?,
-                                            None => {
-                                                error!(
-                                                    "TakeBuy: Buyer pubkey not found for order {}!",
-                                                    order.id
-                                                );
-                                                break;
-                                            }
-                                        };
-                                        // Check market price value in sats - if order was with market price then calculate it and send a DM to buyer
-                                        if order.amount == 0 {
-                                            order.amount = set_market_order_sats_amount(
-                                                &mut order,
-                                                buyer_pubkey,
-                                                &my_keys,
-                                                &pool,
-                                                &client,
-                                            )
-                                            .await?;
-                                        }
-
-                                        show_hold_invoice(
-                                            &pool,
-                                            &client,
-                                            &my_keys,
-                                            None,
-                                            &buyer_pubkey,
-                                            &seller_pubkey,
-                                            &order,
-                                        )
-                                        .await?;
-                                    }
-                                    Action::FiatSent => {
-                                        let order_id = msg.order_id.unwrap();
-                                        let order = match Order::by_id(&pool, order_id).await? {
-                                            Some(order) => order,
-                                            None => {
-                                                error!("FiatSent: Order Id {order_id} not found!");
-                                                break;
-                                            }
-                                        };
-                                        // TODO: send to user a DM with the error
-                                        if order.status != "Active" {
-                                            error!("FiatSent: Order Id {order_id} wrong status");
-                                            break;
-                                        }
-                                        // Check if the pubkey is the buyer
-                                        if Some(event.pubkey.to_bech32()?) != order.buyer_pubkey {
-                                            let text_message = messages::cant_do();
-                                            // We create a Message
-                                            let message = Message::new(
-                                                0,
-                                                Some(order.id),
-                                                Action::CantDo,
-                                                Some(Content::TextMessage(text_message)),
-                                            );
-                                            let message = message.as_json()?;
-                                            send_dm(&client, &my_keys, &event.pubkey, message)
-                                                .await?;
-                                        }
-
-                                        // We publish a new replaceable kind nostr event with the status updated
-                                        // and update on local database the status and new event id
-                                        update_order_event(
-                                            &pool,
-                                            &client,
-                                            &my_keys,
-                                            Status::FiatSent,
-                                            &order,
-                                            None,
-                                        )
-                                        .await?;
-
-                                        let seller_pubkey = match order.seller_pubkey.as_ref() {
-                                            Some(pk) => XOnlyPublicKey::from_bech32(pk)?,
-                                            None => {
-                                                error!(
-                                                    "Seller pubkey not found for order {}!",
-                                                    order.id
-                                                );
-                                                break;
-                                            }
-                                        };
-                                        let peer = Peer::new(event.pubkey.to_bech32()?);
-
-                                        // We create a Message
-                                        let message = Message::new(
-                                            0,
-                                            Some(order.id),
-                                            Action::FiatSent,
-                                            Some(Content::Peer(peer)),
-                                        );
-                                        let message = message.as_json().unwrap();
-                                        send_dm(&client, &my_keys, &seller_pubkey, message).await?;
-                                        // We send a message to buyer to wait
-                                        let peer = Peer::new(seller_pubkey.to_bech32()?);
-
-                                        // We create a Message
-                                        let message = Message::new(
-                                            0,
-                                            Some(order.id),
-                                            Action::FiatSent,
-                                            Some(Content::Peer(peer)),
-                                        );
-                                        let message = message.as_json()?;
-                                        send_dm(&client, &my_keys, &event.pubkey, message).await?;
-                                    }
-                                    Action::Release => {
-                                        let order_id = msg.order_id.unwrap();
-                                        let order = match Order::by_id(&pool, order_id).await? {
-                                            Some(order) => order,
-                                            None => {
-                                                error!("Release: Order Id {order_id} not found!");
-                                                break;
-                                            }
-                                        };
-                                        let seller_pubkey = event.pubkey;
-                                        if Some(seller_pubkey.to_bech32()?) != order.seller_pubkey {
-                                            let text_message = messages::cant_do();
-                                            // We create a Message
-                                            let message = Message::new(
-                                                0,
-                                                Some(order.id),
-                                                Action::CantDo,
-                                                Some(Content::TextMessage(text_message)),
-                                            );
-                                            let message = message.as_json()?;
-                                            send_dm(&client, &my_keys, &event.pubkey, message)
-                                                .await?;
-                                        }
-
-                                        if order.preimage.is_none() {
-                                            break;
-                                        }
-                                        let preimage = order.preimage.as_ref().unwrap();
-                                        ln_client.settle_hold_invoice(preimage).await?;
-                                        info!("Release: Order Id {}: Released sats", &order.id);
-                                        // We publish a new replaceable kind nostr event with the status updated
-                                        // and update on local database the status and new event id
-                                        update_order_event(
-                                            &pool,
-                                            &client,
-                                            &my_keys,
-                                            Status::SettledHoldInvoice,
-                                            &order,
-                                            None,
-                                        )
-                                        .await?;
-
-                                        // Finally we try to pay buyer's invoice
-                                        let payment_request =
-                                            order.buyer_invoice.as_ref().unwrap().to_string();
-                                        let mut ln_client_payment =
-                                            lightning::LndConnector::new().await;
-                                        let (tx, mut rx) = channel(100);
-                                        let payment_task = {
-                                            async move {
-                                                ln_client_payment
-                                                    .send_payment(
-                                                        &payment_request,
-                                                        order.amount,
-                                                        tx,
-                                                    )
-                                                    .await;
-                                            }
-                                        };
-                                        tokio::spawn(payment_task);
-                                        let payment = {
-                                            async move {
-                                                // We redeclare vars to use inside this block
-                                                let client = util::connect_nostr().await.unwrap();
-                                                let my_keys = util::get_keys().unwrap();
-                                                let buyer_pubkey = XOnlyPublicKey::from_bech32(
-                                                    order.buyer_pubkey.as_ref().unwrap(),
-                                                )
-                                                .unwrap();
-                                                let pool = db::connect().await.unwrap();
-                                                // Receiving msgs from send_payment()
-                                                while let Some(msg) = rx.recv().await {
-                                                    if let Some(status) =
-                                                        PaymentStatus::from_i32(msg.payment.status)
-                                                    {
-                                                        if status == PaymentStatus::Succeeded {
-                                                            info!(
-                                                                "Release: Order Id {}: Invoice with hash: {} paid!",
-                                                                order.id,
-                                                                msg.payment.payment_hash
-                                                            );
-                                                            // Purchase completed message to buyer
-                                                            let message = Message::new(
-                                                                0,
-                                                                Some(order.id),
-                                                                Action::PurchaseCompleted,
-                                                                None,
-                                                            );
-                                                            let message =
-                                                                message.as_json().unwrap();
-                                                            send_dm(
-                                                                &client,
-                                                                &my_keys,
-                                                                &buyer_pubkey,
-                                                                message,
-                                                            )
-                                                            .await
-                                                            .unwrap();
-                                                            let status = Status::Success;
-                                                            // We publish a new replaceable kind nostr event with the status updated
-                                                            // and update on local database the status and new event id
-                                                            update_order_event(
-                                                                &pool, &client, &my_keys, status,
-                                                                &order, None,
-                                                            )
-                                                            .await
-                                                            .unwrap();
-                                                        }
-                                                    }
-                                                }
-                                            }
-                                        };
-                                        tokio::spawn(payment);
-                                    }
-                                    Action::Cancel => {
-                                        let order_id = msg.order_id.unwrap();
-                                        let order = match Order::by_id(&pool, order_id).await? {
-                                            Some(order) => order,
-                                            None => {
-                                                error!("Cancel: Order Id {order_id} not found!");
-                                                break;
-                                            }
-                                        };
-                                        if order.status == "Pending" {
-                                            // Validates if this user is the order creator
-                                            let user_pubkey = event.pubkey.to_bech32()?;
-                                            if user_pubkey != order.creator_pubkey {
-                                                let text_message = messages::cant_do();
-                                                // We create a Message
-                                                let message = Message::new(
-                                                    0,
-                                                    Some(order.id),
-                                                    Action::CantDo,
-                                                    Some(Content::TextMessage(text_message)),
-                                                );
-                                                let message = message.as_json()?;
-                                                send_dm(&client, &my_keys, &event.pubkey, message)
-                                                    .await?;
-                                                break;
-                                            }
-                                            // We publish a new replaceable kind nostr event with the status updated
-                                            // and update on local database the status and new event id
-                                            update_order_event(
-                                                &pool,
-                                                &client,
-                                                &my_keys,
-                                                Status::Canceled,
-                                                &order,
-                                                None,
-                                            )
-                                            .await?;
-                                            // We create a Message for cancel
-                                            let message = Message::new(
-                                                0,
-                                                Some(order.id),
-                                                Action::Cancel,
-                                                None,
-                                            );
-                                            let message = message.as_json()?;
-                                            send_dm(&client, &my_keys, &event.pubkey, message)
-                                                .await?;
-                                        } else if order.status == "WaitingBuyerInvoice" {
-                                            // We return funds to seller
-                                            let hash = order.hash.as_ref().unwrap();
-                                            ln_client.cancel_hold_invoice(hash).await?;
-                                            info!(
-                                                "Cancel: Order Id {}: Funds returned to seller",
-                                                &order.id
-                                            );
-                                            let creator = event.pubkey.to_bech32()?;
-                                            if &creator == order.buyer_pubkey.as_ref().unwrap() {
-                                                // We publish a new replaceable kind nostr event with the status updated
-                                                // and update on local database the status and new event id
-                                                update_order_event(
-                                                    &pool,
-                                                    &client,
-                                                    &my_keys,
-                                                    Status::Canceled,
-                                                    &order,
-                                                    None,
-                                                )
-                                                .await?;
-                                                // We create a Message for cancel
-                                                let message = Message::new(
-                                                    0,
-                                                    Some(order.id),
-                                                    Action::Cancel,
-                                                    None,
-                                                );
-                                                let message = message.as_json()?;
-                                                send_dm(&client, &my_keys, &event.pubkey, message)
-                                                    .await?;
-                                            } else {
-                                                // We re-publish the event with Pending status
-                                                // and update on local database
-                                                let mut amount = order.amount;
-                                                let mut fee = order.fee;
-                                                if order.price_from_api {
-                                                    amount = 0;
-                                                    fee = 0;
-                                                }
-                                                update_order_to_initial_state(
-                                                    &pool, order.id, amount, fee,
-                                                )
-                                                .await?;
-                                                update_order_event(
-                                                    &pool,
-                                                    &client,
-                                                    &my_keys,
-                                                    Status::Pending,
-                                                    &order,
-                                                    None,
-                                                )
-                                                .await?;
-                                                info!(
-                                                    "Buyer: {}: Canceled order Id {} republishing order",
-                                                    order.buyer_pubkey.as_ref().unwrap(),
-                                                    &order.id
-                                                );
-                                            }
-                                        } else if order.status == "WaitingPayment" {
-                                            // TODO
-                                            unimplemented!()
-                                        } else if order.status == "Active"
-                                            || order.status == "FiatSent"
-                                            || order.status == "Dispute"
-                                        {
-                                            // TODO
-                                            unimplemented!()
-                                        } else {
-                                            // TODO
-                                            unimplemented!()
-                                        }
-                                    }
-                                    Action::AddInvoice => {
-                                        // Safe unwrap as we verified the message
-                                        let order_id = msg.order_id.unwrap();
-                                        let order = match Order::by_id(&pool, order_id).await? {
-                                            Some(order) => order,
-                                            None => {
-                                                error!(
-                                                    "AddInvoice: Order Id {order_id} not found!"
-                                                );
-                                                break;
-                                            }
-                                        };
-                                        if order.kind != "Buy" {
-                                            error!("AddInvoice: Order Id {order_id} wrong kind");
-                                            break;
-                                        }
-                                        let buyer_pubkey = event.pubkey;
-                                        let pr: String;
-                                        // If a buyer sent me a lightning invoice we look on db an order with
-                                        // that order id and save the buyer pubkey and invoice fields
-                                        if let Some(payment_request) = msg.get_payment_request() {
-                                            // Verify if invoice is valid
-                                            match is_valid_invoice(
-                                                &payment_request,
-                                                Some(order.amount as u64),
-                                            ) {
-                                                Ok(_) => {}
-                                                Err(e) => match e {
-                                                    MostroError::ParsingInvoiceError
-                                                    | MostroError::InvoiceExpiredError
-                                                    | MostroError::MinExpirationTimeError
-                                                    | MostroError::WrongAmountError
-                                                    | MostroError::MinAmountError => {
-                                                        // We create a Message
-                                                        let message = Message::new(
-                                                            0,
-                                                            Some(order.id),
-                                                            Action::CantDo,
-                                                            Some(Content::TextMessage(
-                                                                e.to_string(),
-                                                            )),
-                                                        );
-                                                        let message = message.as_json()?;
-                                                        send_dm(
-                                                            &client,
-                                                            &my_keys,
-                                                            &buyer_pubkey,
-                                                            message,
-                                                        )
-                                                        .await?;
-                                                        error!("{e}");
-                                                        break;
-                                                    }
-                                                    _ => {}
-                                                },
-                                            }
-                                            pr = payment_request;
-                                        } else {
-                                            error!("AddInvoice: Order Id {order_id} wrong get_payment_request");
-                                            break;
-                                        }
-
-                                        let order_status = match Status::from_str(&order.status) {
-                                            Ok(s) => s,
-                                            Err(e) => {
-                                                error!("AddInvoice: Order Id {order_id} wrong status: {e:?}");
-                                                break;
-                                            }
-                                        };
-                                        // Buyer can take pending orders only
-                                        match order_status {
-                                            Status::WaitingBuyerInvoice => {}
-                                            _ => {
-                                                // We create a Message
-                                                let message = Message::new(
-                                                    0,
-                                                    Some(order.id),
-                                                    Action::CantDo,
-                                                    Some(Content::TextMessage(format!(
-                                                        "Order Id {order_id} status must be WaitingBuyerInvoice!"
-                                                    ))),
-                                                );
-                                                let message = message.as_json()?;
-                                                send_dm(&client, &my_keys, &buyer_pubkey, message)
-                                                    .await?;
-                                                break;
-                                            }
-                                        }
-                                        // We send this data related to the order to the parties
-                                        let order_data = SmallOrder::new(
-                                            order.id,
-                                            order.amount,
-                                            order.fiat_code.clone(),
-                                            order.fiat_amount,
-                                            pr,
-                                            order.premium,
-                                            order.buyer_pubkey.as_ref().cloned(),
-                                            order.seller_pubkey.as_ref().cloned(),
-                                        );
-                                        // We send a confirmation message to seller
-                                        let message = Message::new(
-                                            0,
-                                            Some(order.id),
-                                            Action::BuyerTookOrder,
-                                            Some(Content::SmallOrder(order_data.clone())),
-                                        );
-                                        let message = message.as_json().unwrap();
-                                        let seller_pubkey =
-                                            order.seller_pubkey.as_ref().cloned().unwrap();
-                                        let seller_pubkey =
-                                            XOnlyPublicKey::from_str(&seller_pubkey).unwrap();
-                                        send_dm(&client, &my_keys, &seller_pubkey, message)
-                                            .await
-                                            .unwrap();
-                                        // We send a message to buyer saying seller paid
-                                        let message = Message::new(
-                                            0,
-                                            Some(order.id),
-                                            Action::HoldInvoicePaymentAccepted,
-                                            Some(Content::SmallOrder(order_data)),
-                                        );
-                                        let message = message.as_json().unwrap();
-                                        send_dm(&client, &my_keys, &buyer_pubkey, message)
-                                            .await
-                                            .unwrap();
-
-                                        // We publish a new replaceable kind nostr event with the status updated
-                                        // and update on local database the status and new event id
-                                        crate::util::update_order_event(
-                                            &pool,
-                                            &client,
-                                            &my_keys,
-                                            Status::Active,
-                                            &order,
-                                            None,
-                                        )
-                                        .await
-                                        .unwrap();
-                                    }
-                                    Action::PayInvoice => todo!(),
-                                    _ => todo!(),
-                                }
-                            }
-                        }
-                    };
-                }
-            }
-        }
-    }
+    run(my_keys, client, &mut ln_client, pool)
 }
 
 #[cfg(test)]
